requirements.txt
numpy>=1.26
pandas>=2.2
tqdm>=4.66
Pillow>=10.3
matplotlib>=3.8
seaborn>=0.13
scikit-learn>=1.4
opencv-python>=4.9

# Deep learning
torch>=2.2
torchvision>=0.17

